Remove un-unused comments. Change folder name. Move sql entities to sql folder
