+++ src/main/java/Entities/nosql/BookGraphEntity.java
@@ -1,21 +1,11 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Entities.Graph;
+package Entities.nosql;
 
-import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.io.Serializable;
 
-
-/**
- *
- * @author abj
- */
 @Entity
 
 public class BookGraphEntity implements Serializable {
