+++ src/main/java/Entities/sql/Book.java
@@ -1,0 +1,15 @@
+package Entities.sql;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import java.io.Serializable;
+
+@Entity
+public class Book implements Serializable {
+    @Id
+    @GeneratedValue
+    private Integer id;
+    
+    
+}
