+++ src/main/java/Entities/sql/GeoLocation.java
@@ -1,0 +1,14 @@
+
+package Entities.sql;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import java.io.Serializable;
+
+public class GeoLocation implements Serializable {
+    @Id
+    @GeneratedValue
+    private Integer id;
+    
+    
+}
